Add repo counter to button
This is a basic example of an idiomatic ES6 ajax call. Native
XMLHttpRequests seem to work just fine for now; vue-resource may not be
necessary. Note the syntax for methods and the use of fat-arrow binding
of `this`. Javascript is actually becoming somewhat pleasant to write in
with these new rules.